backend/requirements.txt
mediapipe==0.10.9
opencv-python==4.9.0.80
websockets==12.0
python-rtmidi==1.5.8
pyfluidsynth==1.3.3
numpy==1.26.4
